+++ src/main/java/com/example/pro/repository/PatientRepository.java
@@ -1,6 +1,8 @@
 package com.example.pro.repository;
 
 import com.example.pro.entity.Patient;
+import jakarta.persistence.LockModeType;
+import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,4 +13,6 @@ import java.util.Optional;
 public interface PatientRepository extends CrudRepository<Patient, Integer> {
     List<Patient> findAll();
     Optional<Patient> findById(int integer);
+    @Lock(LockModeType.PESSIMISTIC_WRITE)
+    Optional<Patient> findForUpdateById(int integer);
 }